requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==11.0.3
asyncpg==0.29.0
redis==5.0.1
httpx==0.25.2
pydantic==1.10.13
prometheus-client==0.19.0
psutil==5.9.6
GPUtil==1.4.0
python-multipart==0.0.6
Jinja2==3.1.2
aiofiles==23.2.1
sqlalchemy==2.0.23
alembic==1.13.0
python-dotenv==1.0.0
uuid==1.30
numpy<2.0